Api Changes - v3.1.0
Cleaned up the code, to behave more abstract, due to changes and solid. Added better support for V3, as in fixed the DB, that would crash, because the DB wouldn't close it's connection.

+++ TQC-Development/ApiEmily/Repositories/ClanPlatformRepository.cs
@@ -48,7 +48,7 @@ namespace ApiEmily.Repositories
                                     new ClanPlatform(
                                         r.GetString(1),
                                         r.GetString(2),
-                                        (uint)r.GetInt32(0))
+                                        r.GetInt32(0))
                             );
                             }
                             catch (NotSupportedException) { throw new NotSupportedException($"Collection was read-only, an error has occurred."); }
@@ -99,7 +99,7 @@ namespace ApiEmily.Repositories
                             clanPlatform = new ClanPlatform(
                                r.GetString(1),
                                r.GetString(2),
-                               (uint)r.GetInt32(0));
+                               r.GetInt32(0));
                         }
                     }
                     catch (DbException) { throw new ExternalException($"Error occurred while executing command text.") as DbException; }
